% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mood.R
\name{getMOODindices}
\alias{getMOODindices}
\title{Given a numeric matrix, compute the MOOD indices}
\usage{
getMOODindices(data, n = ncol(data), nGroups = n/getOption("mc.cores", 1),
  method = c("rcpp", "classic", "normal", "old"), benchmark = c(1, 0, 1),
  parClus)
}
\arguments{
\item{data}{The numeric matrix from whose columns the indices will be computed}

\item{method}{The MOOD implementation to use}
}
\value{
A data.frame with the MOOD indices, namely \code{shape}, \code{magnitude}, \code{amplitude}
}
\description{
Given a numeric matrix, compute the MOOD indices
}
\examples{
data(mtcars)

my.data <- as.matrix(mtcars)
indices <- getMOODindices(t(mtcars))
tail(sapply(indices, sort)) # possible outliers
}

