% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muod.R
\name{getMUODindices}
\alias{getMUODindices}
\title{Given a numeric matrix, compute the MUOD indices}
\usage{
getMUODindices(data, n = ncol(data), nGroups = n/getOption("mc.cores", 1),
  method = c("rcpp", "classic", "normal", "old"), benchmark = c(1, 0, 1),
  parClus)
}
\arguments{
\item{data}{The numeric matrix from whose columns the indices will be computed}

\item{method}{The MUOD implementation to use}
}
\value{
A data.frame with the MUOD indices, namely \code{shape}, \code{magnitude}, \code{amplitude}
}
\description{
Given a numeric matrix, compute the MUOD indices
}
\examples{
data(mtcars)

my.data <- as.matrix(mtcars)
indices <- getMUODindices(t(mtcars))
tail(sapply(indices, sort)) # possible outliers
}

