% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mood.R
\name{getOutliers}
\alias{getOutliers}
\title{Compute multidimensional outliers on a numeric matrix}
\usage{
getOutliers(data, n = ncol(data), nGroups = n/getOption("mc.cores", 1),
  method = c("rcpp", "classic", "normal", "old"),
  outl.method = c("deriv.old", "deriv-enh", "deriv", "roc", "tangent"),
  slope = 2, benchmark = c(1, 0, 1), plotCutoff = F)
}
\arguments{
\item{data}{A numeric matrix to detect outliers on its columns}

\item{method}{The method to compute the MOOD indices}

\item{outl.method}{The method to choose the outliers among the indices}

\item{slope}{The slope threshold when \code{outl.method = "deriv*"}}
}
\value{
A list of outliers, namely \code{shape}, \code{magnitude}, \code{amplitude}
}
\description{
Compute multidimensional outliers on a numeric matrix
}
\examples{
data(mtcars)

my.data <- as.matrix(mtcars)
outliers <- getOutliers(t(my.data), slope=1)
outliers
}

